feat: carousel view 필요없는 파일 및 코드 삭제
